`FileHandler::returnBytes()`에서 잘못된 값을 반환하는 문제 수정

+++ classes/file/FileHandler.class.php
@@ -636,7 +636,8 @@ class FileHandler
 	function returnBytes($val)
 	{
 		$unit = strtoupper(substr($val, -1));
-		$val = (int)$val;
+		$val = (float)$val;
+
 		switch ($unit)
 		{
 			case 'G': $val *= 1024;
